Split actions and reducers; move to subdir
