+++ src/redux/categories-reducers.js
@@ -1,5 +1,5 @@
-import { store, installReducers } from './shop-redux-store.js';
-import { findCategory, findCategoryIndex } from './shop-redux-helpers.js';
+import { store, installReducers } from './store.js';
+import { findCategoryIndex } from './helpers.js';
 
 installReducers({
   // Init categories
@@ -71,42 +71,3 @@ function initCategories() {
   });
 }
 initCategories();
-
-export function changeOffline(offline) {
-  return (dispatch, getState) => {
-    dispatch({
-      type: '_offlineChanged',
-      offline
-    });
-    if (!offline) {
-      tryReconnect()(dispatch, getState);
-    }
-  };
-}
-
-export function tryReconnect() {
-  return (dispatch, getState) => {
-    const state = getState();
-    const category = findCategory(state.categories, state.categoryName);
-    loadCategory(category, dispatch);
-  };
-}
-
-export function loadCategory(category, dispatch) {
-  if (category) {
-    const categoryName = category.name;
-    if (!category.items) {
-      fetch(`data/${categoryName}.json`)
-      .then(res => res.json())
-      .then(data => dispatch({
-        type: '_categoryItemsChanged',
-        categoryName,
-        data
-      }))
-      .catch(() => dispatch({
-        type: '_failureChanged',
-        failure: true
-      }));
-    }
-  }
-}
