+++ src/redux/store.js
@@ -1,7 +1,7 @@
-import createStore from '../node_modules/@0xcda7a/redux-es6/es/createStore.js';
-import applyMiddleware from '../node_modules/@0xcda7a/redux-es6/es/applyMiddleware.js';
-import origCompose from '../node_modules/@0xcda7a/redux-es6/es/compose.js';
-import thunk from '../node_modules/redux-thunk/es/index.js';
+import createStore from '../../node_modules/@0xcda7a/redux-es6/es/createStore.js';
+import applyMiddleware from '../../node_modules/@0xcda7a/redux-es6/es/applyMiddleware.js';
+import origCompose from '../../node_modules/@0xcda7a/redux-es6/es/compose.js';
+import thunk from '../../node_modules/redux-thunk/es/index.js';
 
 const compose = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || origCompose;
